% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_srs.R
\name{generate_srs}
\alias{generate_srs}
\title{Generate a simple random spatial sample from a polygon layer}
\usage{
generate_srs(path, layer, n, site_code, year, crs_in = 3338, crs_out = 4326)
}
\arguments{
\item{path}{Path to directory containing the shapefile}

\item{layer}{Layer name (string)}

\item{n}{Number of sample points}

\item{site_code}{Code used for naming samples (e.g., "UGOH")}

\item{year}{Year for naming samples}

\item{crs_in}{Input CRS (default = EPSG:3338)}

\item{crs_out}{Output CRS (default = EPSG:4326)}
}
\value{
A tibble with columns Site, Latitude, Longitude
}
\description{
Generate a simple random spatial sample from a polygon layer
}
